requirements.txt
click==8.0.1
decorator==4.4.2
ffmpeg-python==0.2.0
ImageHash==4.2.0
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.0
pytesseract==0.3.7
python-dateutil==2.8.1
python-libxdo==0.1.2a1
PyWavelets==1.1.1
scipy==1.6.1
SQLAlchemy==1.3.23
videoio==0.2.2
git+git://github.com/chalant/python-libxdo.git@master#egg=python-libxdo
